+++ lib/securial/auth/token_generator.rb
@@ -15,6 +15,12 @@ module Securial
 
           "#{hmac}#{random_data}"
         end
+
+        def generate_password_reset_token
+          chars = ("A".."Z").to_a + ("a".."z").to_a + ("0".."9").to_a
+          token = Array.new(12) { chars.sample }.join
+          "#{token[0, 6]}-#{token[6, 6]}"
+        end
       end
     end
   end